Update Graphing_Calculator.ino
I found the problem with parentheses not being popped correctly in the postfixconversion method.  There's still the occasional problem with some sort of syntax on some things, but things like 5+(((2))) work correctly now.  Also changed some comments surrounding the drawgraph method and got rid of the random multiplication I had in it.